requirements.txt
Babel==2.6.0
certifi==2018.11.29
chardet==3.0.4
construct==2.5.3
Django==1.11.17
django-haystack==2.8.1
django-phonenumber-field==2.1.0
djangorestframework==3.9.1
future==0.17.1
haystack==0.42
idna==2.8
MySQL-python==1.2.5
oauthlib==3.0.1
pefile==2018.8.8
phonenumbers==8.10.4
Pillow==6.0.0
PyJWT==1.7.1
pysolr==3.8.1
python-openid==2.2.5
python-ptrace==0.9.3
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
urllib3==1.24.1
